+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+ - Added `session.Database()` which returns the current default database
+
 ## v2.1.0 - 2016-06-26
 
 ### Added